% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMETAR-package.R
\docType{data}
\name{airports}
\alias{airports}
\title{A data frame with 55,345 airports.}
\format{A data frame with 55,345 rows and 17 variables.  See Details.}
\source{
\url{http://ourairports.com/data}
}
\usage{
airports
}
\description{
A dataset containing airport information, downloaded using the
\code{\link[=download_airports]{download_airports()}} function. This is a snapshot of August, 2019.
}
\details{
A data.frame with columns:
\describe{
\item{id}{Airport id}
\item{iata}{IATA code}
\item{continent}{Continent code}
\item{country}{Country code}
\item{country_name}{Country name}
\item{region}{Region code}
\item{region_name}{Region name}
\item{municipality}{Municipality name}
\item{name}{Airport name}
\item{type}{Airport type, e.g. small airport, heliport, etc.}
\item{latitude}{(in degrees)}
\item{longitude}{(in degrees)}
\item{elevation}{(Elevation in feet)}
\item{gps}{GPS code}
\item{local_code}{Local airport code. This can be used to join in the METAR.}
\item{link}{URL}
\item{link_wikipedia}{Wikipedia URL}
}
}
\keyword{datasets}
