% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-data.R
\name{download_airports}
\alias{download_airports}
\title{Fetch a list of airports}
\usage{
download_airports()
}
\value{
a data.frame with columns id, continent, country, country_name, region, region_name, municipality, name, type, latitude (in degrees), longitude (in degrees), elevation (in feet), gps, iata, local_code, link, link_wikipedia
}
\description{
Fetch a list of airports from \url{http://ourairports.com/data}. This data is Public Domain.
}
\examples{
\dontrun{
x <- download_airports()
}
}
