% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-data.R
\name{download_current_metar_cycle}
\alias{download_current_metar_cycle}
\title{Retrieve current METAR data from NOAA}
\usage{
download_current_metar_cycle(hour)
}
\arguments{
\item{hour}{an hour, either one of 0, 1, 2, up to 23}
}
\value{
a data.frame with columns time and metar containing the METARS from the selected hour currently at \url{https://tgftp.nws.noaa.gov/data/observations/metar/cycles}
}
\description{
Retrieve hourly cycles of METAR data from \url{https://tgftp.nws.noaa.gov/data/observations/metar/cycles}
}
\examples{
x <- download_current_metar_cycle(hour = 0)
x <- download_current_metar_cycle(hour = 7)
x <- download_current_metar_cycle(hour = 12)
x <- download_current_metar_cycle(hour = 16)
x <- download_current_metar_cycle(hour = 23)
\dontrun{
x <- lapply(0:23, download_current_metar_cycle)
}
}
