% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-data.R
\name{download_metar}
\alias{download_metar}
\title{Retrieve current METAR data from NOAA.}
\usage{
download_metar(hour)
}
\arguments{
\item{hour}{an hour, either one of 0, 1, 2, up to 23}
}
\value{
a \code{data.frame} with columns \code{time} and \code{metar} containing the METARS from the selected hour currently at \url{https://tgftp.nws.noaa.gov/data/observations/metar/cycles}
}
\description{
Retrieve hourly cycles of METAR data from \url{https://tgftp.nws.noaa.gov/data/observations/metar/cycles}
}
\examples{
x <- download_metar(hour = 12)
\dontrun{
x <- download_metar(hour = 0)
x <- download_metar(hour = 7)
x <- download_metar(hour = 12)
x <- download_metar(hour = 16)
x <- download_metar(hour = 23)
x <- lapply(0:23, download_metar)
}
}
\seealso{
Other data functions: \code{\link{download_airports}}
}
\concept{data functions}
