% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metar.R
\name{read_metar}
\alias{read_metar}
\title{Read metar data}
\usage{
read_metar(stations, hours = 48, wa = weather_abbrev, ca = cloud_abbrev)
}
\arguments{
\item{stations}{METAR stations}

\item{hours}{number of hours back, max =120}

\item{wa}{vector of weather abbreviations}

\item{ca}{vector of cloud-cover abbreviations}
}
\value{

}
\description{
Read metar data
}
\examples{
read_metar(stations="KATL")
}
