% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_metar.R
\name{recent_metar}
\alias{recent_metar}
\title{Isolate recent data}
\usage{
recent_metar(x, path)
}
\arguments{
\item{x}{previously imported metar data}

\item{path}{filepath of existing csv}
}
\description{
Isolate recent data
}
