% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMETAR-package.R
\docType{package}
\name{RMETAR-package}
\alias{RMETAR}
\alias{RMETAR-package}
\title{RMETAR: Functionality to extract weather information from METAR reports.}
\description{
Functions to extract weather information from METAR reports
(Aviation Routine Weather Reports). Using the METAR Decoder Software
Package Library (mdsplib) which is provided by the National Weather
Service.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.bnosac.be}
}

}
\author{
\strong{Maintainer}: Jan Wijffels \email{jwijffels@bnosac.be} [copyright holder]

Other contributors:
\itemize{
  \item Eric McCarthy \email{eric@limulus.net} [copyright holder]
}

}
\keyword{internal}
