% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-data.R
\name{download_airports}
\alias{download_airports}
\title{Fetch a list of airports.}
\usage{
download_airports()
}
\value{
a data.frame with columns \code{id}, \code{iata}, \code{continent}, \code{country}, \code{country_name}, \code{region}, \code{region_name}, \code{municipality}, \code{name}, \code{type}, \code{latitude} (in degrees), \code{longitude} (in degrees), \code{elevation} (in feet), \code{gps}, \code{local_code}, \code{link} and \code{link_wikipedia}
}
\description{
Fetch a list of airports from \url{http://ourairports.com/data}. This data is Public Domain.
}
\examples{
\dontrun{
x <- download_airports()
}
}
\seealso{
Other data functions: \code{\link{download_metar}}
}
\concept{data functions}
